Fix too agressive pointer deletion (BREAKING)
There is a chance that pointer field is named exactly the same like e.g
child node. This commit introduces another check - field must be a
string to be treated as pointer (this does not fundamentally prevent
misrecognition, but should tackle most real cases)

BREAKING CHANGE:
Before, there was `ptrs` jq function, that could be passed to
--del-fields flag. This commit deletes `ptrs` func and introduces
separate flag --del-ptrs. Rationale: it is more feasible to do it in
python than jq (both simpler and faster)

Sidenote:
We could implement most of --del-fields funtionality in python